# apps/api/openapi-v2.yml
openapi: 3.0.3
info:
  title: Edge-Cloud Orchestrator API
  description: Production API documentation for the Edge-Cloud Orchestrator Control Plane
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    def-0:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error description
            requestId:
              type: string
              description: Unique request ID
            timestamp:
              type: string
              format: date-time
              description: ISO 8601 timestamp
            details:
              description: Optional error details
            stack:
              type: string
              description: Error stack trace (non-production only)
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
      title: ErrorSchema
...
paths:
...
  /v2/tasks/artifacts:
    post:
      summary: Upload a WASM task artifact
      tags:
        - tasks
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  artifactId:
                    type: string
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /v2/tasks/artifacts/{id}:
    get:
      summary: Download a WASM task artifact
      tags:
        - tasks
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: string
                format: binary
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
  /v2/tasks/{id}:
    get:
      summary: Get task by ID
      tags:
        - tasks
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  type:
                    type: string
                  status:
                    type: string
                    enum:
                      - PENDING
                      - SCHEDULED
                      - RUNNING
                      - COMPLETED
                      - FAILED
                      - CANCELLED
                  priority:
                    type: string
                  nodeId:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                  specs:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: object
                            properties:
                              cpuCores:
                                type: number
                              memoryGB:
                                type: number
                            required:
                              - cpuCores
                              - memoryGB
                            additionalProperties: false
                      - type: "null"
                  submittedAt:
                    type: string
                    format: date-time
                  startedAt:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                            format: date-time
                      - type: "null"
                  completedAt:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                            format: date-time
                      - type: "null"
                  error:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                  metadata:
                    type: object
                    additionalProperties: {}
                  runtime:
                    type: string
                    enum:
                      - NATIVE
                      - DOCKER
                      - WASM
                  image:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                  wasmArtifactId:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                  affinity:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                  traceId:
                    anyOf:
                      - anyOf:
                          - not: {}
                          - type: string
                      - type: "null"
                required:
                  - id
                  - name
                  - type
                  - status
                  - priority
                  - submittedAt
                  - runtime
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /v2/tasks/{id}/cancel:
    post:
      summary: Cancel a task
      tags:
        - tasks
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /v2/tasks/{id}/logs:
    get:
      summary: Get task logs
      tags:
        - tasks
      parameters:
        - schema:
            type: string
          in: query
          name: level
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 1000
          in: query
          name: limit
          required: false
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /v2/tasks/{id}/retry:
    post:
      summary: Retry a failed task
      tags:
        - tasks
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /v2/tasks/{id}/scheduling-decision:
    get:
      summary: Get scheduling decision for a task
      tags:
        - tasks
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...